moved rust client and server to main dir
